requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
decorator==5.1.1
defusedxml==0.7.1
Django==5.0
django-allauth==0.60.0
django-debug-toolbar==4.2.0
django-email-users==1.1.0
django-filter==23.5
django-froala-editor==4.1.4
executing==2.0.1
idna==3.6
ipython==8.19.0
jedi==0.19.1
Markdown==3.5.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
urllib3==2.1.0
wcwidth==0.2.13
